modbus: make iface an signed integer
The Modbus API defines the interface as a signed integer.

However, there is an inconsistency in the Modbus API with the function
modbus_disable() that use an unsigned integer.

This updates the modbus_disable() prototype to use a signed integer iface
and applies the change in the direct subsequent call core function
modbus_get_context() and in the indirect functions modbus_init_server() and
modbus_init_client().

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>